+++ dotnet/Services/IVtexAPIService.cs
@@ -13,8 +13,8 @@ namespace ShipStation.Services
         Task<OrderInvoiceNotificationResponse> OrderInvoiceNotification(string orderId, OrderInvoiceNotificationRequest orderInvoice);
         Task<bool> SetOrderStatus(string orderId, string orderStatus);
         Task<bool> ProcessShipNotification(ListShipmentsResponse shipmentsResponse);
-        Task<ListAllDocksResponse> ListAllDocks();
-        Task<ListAllWarehousesResponse> ListAllWarehouses();
+        Task<ListAllDocksResponse[]> ListAllDocks();
+        Task<ListAllWarehousesResponse[]> ListAllWarehouses();
         Task<string> ValidateShipments(string date);
         Task<string> ValidateOrderShipments(string orderId);
         Task<string> CheckCancelledOrders(DateTime date);